fix enum sentinel prop tests
Summary:
Back in D15383777, I removed a case because it was causing the following code to fail:

```
declare var Enum: {|
  A: 'A',
  B: 'B',
|}

function foo(e: 'A' | 'B') {
    if (e === Enum.A) {
    }
}
```

producing a spurious incompatibility between `A` and `A`. I removed the enum case that handled this code, but didn't realize that this allowed the following unsound code:

```
type T = 'A' | 'B';

const E = {A: 'A', B: 'B'};

declare var x: T;

if (x === E.A) {
  (E.A: T); // OK
  (E.A: number); // Error expected
}
```

It turns out we still want to trigger these flows for all the non-matching cases, so we re-add the enum upper bound case for these scenarios.

Reviewed By: panagosg7

Differential Revision: D21232016

fbshipit-source-id: 8bbc16a7f91f54adfa0f35dbf5ac511a17bd3fe1