Do not add blank lines around JSX when printing new expression arguments
Summary:
We currently have a formatting issue when printing JSX arguments inside new expressions. When mutliline JSX is an argument of a new expression, prettier will wrap and indent it on a new line in parens. However we do not have the loc of parens when determining whether to preserve blank lines before and after an argument, as the loc of parens is not preserved in the AST. This means that from the perspective of the `js_layout_generator` this multline JSX argument has a blank line before and after it, as its loc is separated from its neighboring arguments by more than one line. The result of this is that currently an extra blank line is inserted before and after multiline JSX in new epxressions.

We can fix this by detecting when we are in this case, and then expanding the loc of the JSX by one line in both directions to account for where the parentheses are. This is pretty hacky and not a perfect solution, but it stops those extra blank lines from being inserted while printing, which was a common problem when reprinting files.

Reviewed By: mroch

Differential Revision: D21266258

fbshipit-source-id: e8c6c5c89b4c91a763f2cc49e0f621b3f03a5aca