Object bracket spacing option for js_layout_generator
Summary:
Prettier has a `bracket_spacing` option which controls whether a space should be inserted between the brackets and properties of an object. Currently the `js_layout_generator` always outputs object literals with a space between the brackets and prop.

Since this is an option in prettier, let's make it an option here as well for accurate reprinting as it may differ between codebases. We can add it to the `js_layout_generator` options and have it default to true, to match the current behavior of the `js_layout_generator`. Let's then disable it for codemods.

Reviewed By: mroch

Differential Revision: D21264015

fbshipit-source-id: 047c92e348f1973d9844a37ac022e82e232d0980