Fix printing of leading comments within computed member expressions
Summary:
When printing computed member expressions, we are printing leading comments on the property on the same line as the `[`.

We can fix this by detecting when the computed property has leading comments, and if so we should break and indent so that all leading comments are printed on the line after the `[` with proper indentation. Now we can accurately reprint the following:

```
foo[
  // L
  a
]
```

Reviewed By: pieterv

Differential Revision: D21220114

fbshipit-source-id: f27ddd9cdab6e93c8295001dcf69e2525b731d6f