[Likely to cause errors] Encode object type invariants into the ocaml representation
Summary:
We have several kinds of objects in Flow:
* Exact
* Inexact
* Objects with indexers (they don't really fit into exact or inexact, even though they use inexact syntax today)
* Unsealed

To express these types, we have the following fields in ocaml:
* dict_t
* flags.sealed
* flags.exact

The way those fields combine to produce our different kinds of objects is:
* Exact: flags.exact = true, flags.sealed = sealed, dict_t = None
* Inexact: flags.exact = false, flags.sealed = sealed, dict_t = None
* Indexed: flags.exact = false, flags.sealed = sealed, dict_t = Some dict
* Unsealed: flags.exact = true, flags.sealed = UnsealedInFile, dict_t = None

Notice that there are still several combinations of these fields not enumerated above-- each of those are object types that we do not have a good intuition for and behave weirdly.

In practice, it's possible to create these kinds of objects:
```
type ExactWithIndexer = {| [string]: number |};
declare var a: ExactWithIndexer;
const unsealedWithIndexer = {...a, ...{}};
declare var b: {};
const unsealedInexact = {...b, ...{}};
```

To make these invariant violations impossible, we can encode the four object kinds we expect directly into the OCaml type system. This diff does exactly that, with the goal of preserving the existing behavior of the types as much as possible. Errors are caused by exact objects with indexers.

## $Exact and {| ...spreads  |}
Since we cannot have things like exact objects with indexers now (not that they ever made sense), the behavior of `$Exact` and spreads with `{||}` was changed to support existing code. `$Exact` will only transform Inexact objects to Exact objects and otherwise do nothing. Spreads with `{||}` will allow objects with indexers to be spread inside them.

See indexed_objects tests for examples of this behavior.

## Fixes

* Allows "exact object types with indexers" syntax to be inhabited:
```
const x: {| [string]: number |} = {foo: 3};
```

This makes most sense when using exact-by-default syntax:
```
const x: { [string]: number } = {foo: 3};
```

* Fixes an issue with spread widening. Previously, we assumed that exact object types did not have indexers. Consider a case where the first object type we see is `{| [string]: number |}` and the second we see is `{| foo: number |}`. Our widening would see that the second object didn't have an indexer, so it would remove the indexer from the widest object type. When indexers are only present on inexact objects, that makes the new type we produce there inexact without an indexer. However, since that invariant was not true, we produced too-precise types unsoundly by dropping the indexer (`{| foo: number |}`). See the added test in object_widening

Reviewed By: samwgoldman

Differential Revision: D21066473

fbshipit-source-id: 762437e15d71278f6bfe08e6c5cc687ce6c0ce92