Fix precedence of type casts when printing
Summary:
It looks like the `js_layout_generator` is always wrapping type casts in an extra set of parens whenever there is a precedence check, such as on the right hand side of a variable declaration.

```
// Input
const a = (b: mixed)

// Output
const a = ((b: mixed));
```

This is because the `js_layout_generator` is treating type casts as if they have the lowest precedence, causing them to always be wrapped in an extra set of parens when precedence is checked. Instead type casts should be treated as having the highest precedence, so that an extra set of parens is never added. There is no case where an extra set of parens is needed as the type cast already contains a built-in set of parens.

Reviewed By: gkz

Differential Revision: D21487328

fbshipit-source-id: cf479683978fbca5494823f4e6ce75c2ea14a119