Fix types-first signatures for class extends
Summary:
Before this change, the types-first signature generator was only able to
generator signatures for class extends clauses that had one of the following
forms:

1. A simple identifier (optionally with type arguments)
2. A simple member expression (optionally with type arguments)

In any other case, the signature generator would output
`$TEMPORARY$Super$FlowFixMe`, which is treated like `any`.

However, the signature _verifier_, which is responsible for asking for necessary
annotations, accepts any literal expression in this position. This can lead to
issues where the verifier accepts a program, but the generator does not produce
a reasonable signature.

For example, it is relatively common to use a type cast expression in the
extends clause under types-first which do not fall into one of the two forms
above, since the signature verifier would generally ask for an annotation.

This diff improves the signature generator by outlining any literal expression
and using the resulting simple identifier in the class extends clause. This
brings the signature generator up to par with the signature verifier.

Reviewed By: panagosg7

Differential Revision: D20377409

fbshipit-source-id: 84aeacabcc17d051dbc02fd4773404f8a94a85d6