Separate TypeUtils from Type
Summary:
The TypeUtils module defines functions like `reason_of_t` which currently
perform very shallow computations over the structure of types.

These kinds of computations over the structure of types are very limited. In
many cases, a Type.t itself cannot be understood without also looking up
associated information in the context.

However, such computations can not live in the Type module, because that would
introduce a circular dependency. That is, Context already depends on Type, so
Type can not also depend on Context.

This diff extricates TypeUtil from the set of recursive modules in type.ml. This
makes it possible for TypeUtil to depend on Context without introducing a
circular dependency. This will become necessary in a later change, where we need
to interrogate the context from `reason_of_t`.

Reviewed By: dsainati1

Differential Revision: D21529691

fbshipit-source-id: 602d0cf7f309586a2d4742e314b37516266e7f60