[Errors] Add a code for each error
Summary: This assigns codes to errors. The codes were determined by inspecting the error output of our test cases and classifying each error with a code that matched its nature. Note that different messages can result in the same code, or that codes in certain cases are based on the `use_op` rather than the `error_message`. This is to reflect the fact that these internal differences are not discernible in the printed error.

Reviewed By: samwgoldman

Differential Revision: D19760727

fbshipit-source-id: 7028aef5cfe210a585b35aa153c1ca86ec91ffb7