mvitousek reverting 7028aef5cfe210a585b35aa153c1ca86ec91ffb7 | sainati | sainati merging c36e5b21d54b27cffedbb01d03d86b52f8f00fed | sainati | [Flow][Errors] Add a code for each error
Summary: This assigns codes to errors. The codes were determined by inspecting the error output of our test cases and classifying each error with a code that matched its nature. Note that different messages can result in the same code, or that codes in certain cases are based on the `use_op` rather than the `error_message`. This is to reflect the fact that these internal differences are not discernible in the printed error.

Reviewed By: samwgoldman

Differential Revision: D19760727

fbshipit-source-id: c6710b68a30281ba284aeb49b36efdf7399b54b4