improve speculative error codes
Summary:
In general, speculative error messages are structured as follows:

One of `EIncompatible`, `EIncompatibleDefs` and `EUnionSpeculation` serves as the root message, containing the `use_op` that expresses the operation that failed, and a list of branches containing each speculative error.

The original design for speculative error codes was to use information about the `upper_kind` of the root incompatibility to determine the code, but in practice it turns out that the `upper_kind` of the root is often ignored by the printer in favor of the branches. Instead we can base the speculative error on the `use_op` in speculative cases.

Reviewed By: mroch

Differential Revision: D21577805

fbshipit-source-id: 4d205d89f693fd728884a3bf64f7e657c18a8407