Convert program AST node from tuple to record
Summary:
We would like to add additional fields to the root program AST node, such as a field for comments that are attached to the program node itself. However the current program AST node is defined as a tuple of the program's loc, list of statements, and list of all comments:

```
type ('M, 'T) program = 'M * ('M, 'T) Statement.t list * 'M Comment.t list
```

If we were to add yet another field to this tuple it would be pretty confusing and hard to use, since you'd have to keep track of the positions of all 4 field, two of which are comments.

Instead, let's first convert the program AST node to a record (like the vast majority of other AST nodes), so that we can add fields more easily in an upcoming change. This will also make it much easier to make further modifications to the program node in the future. The new signature for the program AST node is:

```
module Program : sig
  type ('M, 'T) t = 'M * ('M, 'T) t'

  and ('M, 'T) t' = {
    statements: ('M, 'T) Statement.t list;
    all_comments: 'M Comment.t list;
  }
end
```

This mostly involved updating all references to `Flow_ast.program` to `Flow_ast.Program.t`, and updating all sites where program nodes are created or destructured.

Reviewed By: mroch

Differential Revision: D21440968

fbshipit-source-id: 1959301b5da4e58464470185aa62df0786cb0d40