Program node has comments for program itself directly attached to AST
Summary:
The comments that are meant to refer to the file itself should not be attached to the first node in the AST. Instead, we should have these program comments attached to the program node itself in the regular `comments` field. This will make comments attached to the program (e.g. the docblock) easier to find. Furthermore, it makes sense to not have the program comments be attached to any node when reprinting/changing files, as changing or moving a node should move its attached comments as well. If the program comments were attached to the first node they would be moved or removed when that first node is changed, unless checked for in every possible codemod.

Let's create a new `Eat.program_comments` function that is called immediately upon parsing a program. This function consumes and returns the initial program comments comments to be attached to the root program node, and then marks them as consumed so they are not attached to the first statement.

Only comments before the first token could potentially be program comments comments. We consider all comments through the last comment with a `flow` directive to be the program comments. If there is no `flow` directive we consider the very first comment the program comment if it starts with `/**`.

Unfortunately due to `js_of_ocaml` we can't use either the shared `String_utils` module or the built in `Str` module for better string searching utility functions. Instead I had to roll my own quick one that simply checks if there is a `flow` directive in a comment and treats it as a program comment if so.

Note that since the `comments` field was added to the `Program` node, we must also update all sites that create or destructure `Program` nodes to take into account this new field.

Reviewed By: pieterv

Differential Revision: D21362364

fbshipit-source-id: 62a1b85911a3ab45fe73a75eabdb3a9fdd7f7a0b