Include attached comments when removing or replacing nodes in AST differ
Summary:
When patches are generated from the AST differ, they use the loc included in the `Delete` or `Replace` `change` to determine which section of the old source file to remove/replace. However this does not include the location of comments that are attached to this node, so the node will be removed from the file but leave its comments around the location it used to be. When writing back a new node we write it with its attached comments, which can lead to creating duplicate comments from the current `Ast_diff_printer`:

```
/* Before */

// Leading
expr;

/* After removal */

// Leading

/* After printing again with comments */

// Leading
// Leading
expr;
```

Instead we should be including the total range of the node and its attached comments when determining which section of the source file to remove/replace. We can use the existing comment bound utilities to expand the node's loc with all of its attached comments, for each type of change node in the AST differ.

Note that this also required add the loc of template literals to  `Flow_ast_differ.node`, as template literal AST nodes do not store their own loc.

Reviewed By: pieterv

Differential Revision: D21362868

fbshipit-source-id: 2f2e1b89a30f5fd3e9f17edf94af946a4e74f272