Improve abnormal control flow handling in logical expressions
Summary:
Improve abnormal control flow handling in logical expressions.

With logical operators the LHS is always evaluated. So if the LHS throws, the whole expression throws. To model this we do not catch abnormal exceptions on the LHS.  As such, we only analyze the RHS expression if the LHS does not throw.  If the LHS does not throw, and the RHS does throw, then we cannot say that the entire expression throws, because we only evaluate the RHS depending on the value of the LHS.  Thus, we catch abnormal control flow exceptions on the RHS and do not rethrow them.

Note that the only kind of abnormal control flow that should be raised from an expression is a Throw. The other kinds (Return, Break, Continue) can only arise from statements, and while statements can appear within expressions (e.g. function expressions), any abnormals will be handled before they get here.

This seems pretty esoteric now, but may become more common if the "throw expressions" proposal becomes reality https://github.com/tc39/proposal-throw-expressions (currently stage 2).

mvitousek previously fixed the handling of `invariant()` in ternary expressions D10414664

Reviewed By: samwgoldman, mvitousek

Differential Revision: D21566904

fbshipit-source-id: 3e22350798fcfdbf0326cbc9c109814785135bb8