Remove manual context lifetime management APIs
Summary:
These functions were necessary because sig contexts included information that
was not intended to persist in the heap. After D21646085, this is no longer a
concern, so the APIs can be removed.

After this change, any data accumulated in the component context will still be
there, and is still strongly referenced from the component context. This might
lead to higher memory usage if something holds a strong reference to the
component context for longer than it needs to.

Reviewed By: mvitousek

Differential Revision: D21646720

fbshipit-source-id: 27b8e8ac9008e7ff0334ec30a8208c3075e87dc1