Factor "component context" from "sig context"
Summary:
The Context module defines a record which is passed around during type checking,
providing access to commonly needed data. The data in this record can be broken
according to their lifetimes into three categories:

1. Per-file information is needed during the AST traversal, to answer questions
like "what options are enabled" where options can be set on a perf-file bases,
like the ability to munge underscores.

2. Per-component information is needed during constraint solving, which happens
outside the context of any specific file -- specifically when dealing with
constraints between files in a cycle.

3. Inter-component information, i.e., stuff that we might want to know about
dependencies, like what modules they export and what types correspond to what
resolved tvars.

Before this change, data for both (2) and (3) were stored in a single record:
Context.sig_t. This means that the lifetime of fields in that record needed to
be manually managed. Specifically, it is crucial to "clear intermediates" before
storing the sig cx into the heap, since that information is no longer needed and
would consume inordinate space.

This change introduces a new "component context" with fields extracted from the
sig context. This new record reflects all the data in category (2), leaving the
sig context with only data in category (3).

Reviewed By: mvitousek

Differential Revision: D21646085

fbshipit-source-id: a63d4dde768001a888c59f634fb2b7fc868bf6fc