[Errors] improve exact incompatibility errors when lower bounds are not objects
Summary: When we have a non-object lower bound flowing to an exact upper bound, we produce a wonky error message along the lines of `inexact number is compatible with exact object`. The wording on this is weird, which by itself is cause enough for this change, but now that we have error codes it also results in `let x : {||} = 3` and `let y :{...} = 3` to have different error codes (`incompatible-exact` and `incompatible-type`) which is undesirable.

Reviewed By: samwgoldman

Differential Revision: D21461744

fbshipit-source-id: 4a2fc848c288658d9bb10ffa53ff1ff44a52c174