pin website build to current docker image
Summary:
the website is built by a docker image that installs a specific version of ruby. using `latest` means we can't upgrade the ruby version because we can't update the docker image and the pinned ruby version atomically.

so, we pin to the current docker image, then we'll publish a new docker image with a newer ruby, then bump Gemfile and config.yml.

Reviewed By: panagosg7

Differential Revision: D21723165

fbshipit-source-id: caca7073442537674157608c3e79dfdd3d9cb3b5