Lint rule for accessing default property of import star
Summary:
Implements the new `default-import-access` lint rule, which errors when the default export of a module is accessed through a module object:

```
import * as Foo from 'Foo';

// Errors
Foo.default;
Foo['default'];
```

The `default-import-access` lint error is gated behind the `experimental.strict_es6_import_export` flowconfig flag to assist with testing and rollout.

To generate these errors we first visit the toplevel statements of a program and gather every `import *` specifier, indexed by their id's loc. Then, we visit every member node and error if the `_object` is a use of one of the `import *` identifiers (determined using the `Scope_api` utilities), and the property is the identifier `default` or the string literal `'default'`.

In the future this will be extended to also catch the default export being accessed via destructuring of an `import *`.

Reviewed By: pieterv

Differential Revision: D21604620

fbshipit-source-id: 2812cbc7aa012f3fcd652d290834ec36e9b2c671