Add infrastructure for upcoming ES6 import/export lint rules
Summary:
We would like to support an upcoming set of rules for ES6 imports/exports (e.g. no `this` in exported functions, or no accessing the `default` property of named imports).

This change sets up the infrastructure for these rules, so that they can be implemented within the new `Strict_es6_import_export` module in upcoming changes.

Let's set up these checks behind the new `experimental.strict_es6_import_export` flowconfig flag. This flag is stored in the options and passed down to the context metadata, so that the upcoming imports/exports checks only run if this experimental flag is set.

The new errors are checked alongside the other post-merge errors in `Merge_js.merge_component`, and simply defer to `Strict_es6_import_export.detect_errors` to detect and add errors to the current context. `Strict_es6_import_export.detect_errors` is currently a stub that performs no work, but will be filled in soon. Note that these lint rules will require an `(ALoc.t, ALoc.t) Ast.Program.t` instead of the typed AST in order to use the existing scope builder utilities, which will be necessary for some rules. Luckily this AST is already available in `Merge_js.merge_component`, we just need to pass it down to the `Strict_es6_import_export` module/

Reviewed By: panagosg7

Differential Revision: D21604598

fbshipit-source-id: b886987fba793e3b3abfebdb486363b154580f35