fix website build's cache keys
Summary:
some of the ruby gems used to deploy flow.org are native extensions, compiled against libc. when I downgraded from Debian 10 to Debian 9, the cached gems stopped working because they expected a newer version of glibc.

to avoid this in the future, this diff adds details about the Debian version to the cache keys. a stable version of Debian won't do a major upgrade to glibc.

Reviewed By: jbrown215

Differential Revision: D21791333

fbshipit-source-id: 3f9f18e3ad6276f225da45e2b21ad25d34ab4060