Lint rule for usage of this in exported functions
Summary:
One rule we would like to be able to enforce for es6 imports/exports is that exported functions cannot contain `this`.

```
// All references to `this` should error
export default function defaultFunc() { this }
export default function() { this }
export default () => { this }

export function named() { this }
export const named = function() { this }
export const named = () => { this }

function named1() { this }
const named2 = function() { this }
const named3 = () => { this }
export {named1, named2, named3}
```

This new lint rule is called `this-in-exported-function`, and is gated behind the `experimental.strict_es6_import_export` flowconfig flag to assist with testing and rollout.

There are two main additions we need to make to implement this lint rule. First, we need a visitor that can check whether a function contains `this` and produces an error at the appropriate loc if so. We can simply run this visitor whenever we encounter a directly exported function declaration, function expression, or arrow functioin in both named and default exports.

The tricker case is when a function is exported using a list of export specifiers (e.g. `function foo() {}; export {foo}`. To handle these cases we now create a map of function AST nodes along with the location of the id they are bound to, which is constructed in the initial pass to gather declarations. Then when we encounter an exported list of ids, we simply look up the definition in this map and run the `this` visitor on the function node if the id is for an exported function.

Note that using `this` within a class is still okay. We also allow using `this` inside a class that is declared within an exported function.

This strategy will not catch functions bound from destructuring declarations or those that are the result of more complex exported expressions, but this rule catches the most common cases.

Reviewed By: panagosg7

Differential Revision: D21675332

fbshipit-source-id: 83ca9ce087b3fda7984f0382e1285fae17af6a80