Lint rule for non-const variable exports tracks variables in export specifier list
Summary:
One rule we would like to be able to enforce for es6 imports/exports is that all exported variables are `const`. We are already checking this for directly exported variable declarations (e.g. `export const x = 1`), but we are not yet checking this for variables that are declared and then exported in a separate statement:

```
var x = 1;

// Should error
export {x};
```

This required gathering a bit more information about declarations in the `Strict_es6_import_export` module. I refactored the initial pass that gathered import star locs into a more general pass that gathers necessary declaration information we need for generating errors. Now we accumulate both the set of import star locs, as well as a map from var decl id locs to the kind of that var decl (const, let, or var).

Then we need to visit the specifier lists in all export declarations. For each locally exported id we look up its definition using the `Scope_api` which gives us its def loc. We can use this to find the var decl's kind using the new map from var decl locs to their kind, to determine which exported ids are for non-const variables.

Reviewed By: panagosg7

Differential Revision: D21653408

fbshipit-source-id: 0647fecbb3d662dac7db566391df92cc8586ff04