remove watchman from types_js
Summary:
currently, the monitor process talks to watchman to get the mergebase and subscribe to changes. the query that gets the mergebase also returns the files changed since then, but we ignore that. simultaneously, the server process talks to watchman to get the files changed since the mergebase, to focus them in lazy mode, but doesn't care about the actual mergebase. the two queries are *identical*. further, if this query fails in either place, we exit. redundant work is bad. having two points of failure is bad.

instead, the monitor should be the only thing talking to watchman, and *it* should notify the server to focus/recheck the changed files, like the monitor would if it received a `force-recheck --focus` command or any other watchman event.

the main difference is that we now treat the changes since mergebase as a recheck instead of part of init. On one hand, you could argue that it's not initialized until these files are checked, but on the other, it should make the server available for other things (including canceling that recheck!) sooner.

note: during init, we parse the whole codebase and calculate deps. then we recheck the files changed since the mergebase, which reparses and recomputes their deps in the process of focusing/checking them. We should avoid that, but it's how it worked before as well, so it's reserved for future work.

Reviewed By: gkz

Differential Revision: D21503420

fbshipit-source-id: 5dc25f021cc5a39a1dd2b15d094f58018cf08aef