[PR] DOCS: Update the repository URL for flow-remove-types in the docs
Summary:
<!--
  If this is a change to library defintions, please include links to relevant documentation.
  If this is a documentation change, please prefix the title with [DOCS].

  If this is neither, ensure you opened a discussion issue and link it in the PR description.
-->

The existing URL 301s to https://github.com/facebookarchive/flow-remove-types, which is archived.

It was actually not obvious that the package was moved into this tree.  I only found it via npm's repository link.  Maybe the description of https://github.com/facebookarchive/flow-remove-types could be updated to point here as well?
Pull Request resolved: https://github.com/facebook/flow/pull/8388

Reviewed By: gkz

Differential Revision: D21814329

Pulled By: mroch

fbshipit-source-id: 16a6fed1f17757e4d8f77d6167f8859af85c460d