[enums] Treat successful exhaustive checks of enums with no default, which exit from each case, as an exit overall (do not implicitly return void)
Summary:
If a switch has a successful exhaustive check (only possible with enums for now), does not have a `default`, and exits (e.g. `return`, `throw`) from every case, then we treat the whole switch as exiting, for the purposes of not adding an implicit `void` return to a function.

```
enum E {
  A,
  B,
}

function f(e: E): string {
  switch (e) {
    case E.A:
      return "Eh";
    case E.B:
      return "Bee";
  }
}
```
Before this diff, we would get an error `string is incompatible with implicitly-returned undefined`. Now, we allow the above example with no errors.

If the switch is not exhaustive (this would be an error, but could be suppressed), then we consider the below example as implicitly returning undefined:
```
function f(e: E): string { // Error: string is incompatible with implicitly-returned undefined
  // $FlowFixMe: non-exhaustive check
  switch (e) {
    case E.A:
      return "Eh";
  }
}
```

How do we do this analysis?

At the top of every function body we add an internal variable (`let` style), named `maybe_exhaustively_checked`. At the end of the function, if we have *not* exited abnormally (e.g. consistent `return` or `throw`) through our existing analysis, we get that internal variable from the environment and flow it to `FunImplicitVoidReturnT`. If at any point during the analysis of the function, some type has flowed to `maybe_exhaustively_checked` (other than `empty`), then we will trigger `FunImplicitVoidReturnT`, and the function will return `void` implicitly.

There are two ways some type can flow to `maybe_exhaustively_checked`:

First, when analyzing `switch` statements we create a tvar called `incomplete_out`, and if there is no `default`, supply it to `EnumExhaustiveCheckT`. We then flow it to `maybe_exhaustively_checked` through `Env.init_let`. During the analysis of the exhaustive checks, if the exhaustive check is not successful (e.g. the discriminant is not an enum, the check has an invalid form, or we haven't checked all cases), then we flow some arbitrary type (e.g. `VoidT`) to `incomplete_out`. This type flows to `incomplete_out`, which flows to `maybe_exhaustively_checked`, which flows to and triggers `FunImplicitVoidReturnT`. If the exhaustive check is succesful these flows don't happen.

Second, how do we deal with branching and other control flow? We use our existing logic for initializing variables to deal with everything. If some branch exits (e.g. `return` or `throw`), we don't need to consider if a variable was initialized in that branch. For all other branches, it is only considered consistently initialized if every non-exiting branch initializes it. Otherwise we flow `void` to that variable. This means that in a function, either every branch needs to exit or end in a switch, or else we flow `void` (due to non-consistent initialization) to `maybe_exhaustively_checked`, which flows to and triggers `FunImplicitVoidReturnT`, which means the function implicitly returns void. If every non-exiting branch does end with a switch, then the logic in the previous point handles that.

E.g. this code
```
function f(e: E): string {
  if (randomBool()) {
    switch (e) {
      case E.A:
        return "Eh";
      case E.B:
        return "Bee";
    }
  } else {
    h();
  }
}
```
Is like this:

```
function f(e: E): string {
  let maybe_exhaustively_checked;
  if (randomBool()) {
    switch (e) {
      case E.A:
        return "Eh";
      case E.B:
        return "Bee";
      default:
        maybe_exhaustively_checked = ...
    }
  } else {
    h();
    // No exit or switch here
  }
}
```
In that case, `maybe_exhaustively_checked` is not consistently initialized, so void flows to it, triggering the implicit void return.

It also neatly handles branches, as the result of initializing in different branches is having `maybe_exhaustively_checked` be the union of each branch. So each branch must exhaustively check succesfully.

There is a limitation due the implementation: if there multiple switches in sequence, we only consider the last one. If the last switch is non-exhaustive, even if there were previous switches that were, we will implicitly return void. I think this is fine because it is safe, and it is an unlikely scenario. The latter switch is actually dead code. At some point we can flag that it is dead code, which would be nice.

In the future I also plan on implementing this for arbitrary switches (even for non-enums), after I launch enums.

Reviewed By: samwgoldman

Differential Revision: D21814721

fbshipit-source-id: 06cb3318079fd83ae548bb34fd46249186fb8ac0