[Errors] standardize suppression comment syntax
Summary:
Prior to this diff, what counted as a valid suppression comment was specified purely within the `.flowconfig` of a project as a regular expression. This is unfortunate, because with upcoming changes adding error codes we had no way to enforce any structure on the suppression, and thus had no idea where to look for an error code.

Instead, we standardize the suppression comment syntax to a simpler version:
`//<SUPPRESSOR>[CODE]`
where `SUPPRESSOR` is one of `$FlowFixMe`, `$FlowIssue` or `$FlowExpectedError`.

Reviewed By: mroch

Differential Revision: D19847216

fbshipit-source-id: 37e609963c0d3a6a5f6ee372c45eed23274038e0