fix restart loop when .flowconfig has changed since hg mergebase
Summary:
when the flow server starts in watchman lazy mode, it parses and computes dependencies but does not focus or check anything.

meanwhile, the monitor talks to watchman and gets the files changed since mergebase. after init, we focus and check those files.

when a client, like `flow status`, sends a request to the server, we check for any watchman notifications we haven't processed yet and send them to the server.

previously, this latter notification got enqueued before the initial notification, so we told the server that the files that were changed since mergebase had actually been changed since we started. that causes the server to restart.

now, we keep track of whether a set of changes are the initial changes since mergebase in the watchman handler, so that *any* monitor -> server message, including `flow status`, will trigger the "initial" check rather than a recheck.

Reviewed By: samwgoldman

Differential Revision: D21847126

fbshipit-source-id: bb83475e57d295280d4af4969cf0fb8c2c7c30b7