[normalizer] Simplify member expansion logic
Summary:
Unlike most normalizer clients, when called from the autocomplete command,
the normalizer attempts to return an object-like structure. Until now, the mode
of operation (member-expanding or not) was determined by information
stored in the environment. Then, at various parts of the normalization process
we would look up the environment to see if we should be expanding the type
to an object structure. The awkward part of this design, was that the logic for
this feature was spread throughout the normalizer module.

This diff moves all the code related to member-expansion to one place, at the
entry point to normalizer. This change was enabled after enforcing a stricter
structure on the normalized types (D20070324).

Regarding the implementation, this diff introduces two modules: TypeConverter,
that holds the code that does normalization in the normal mode, and
ExpandMembersConverter, that handles the case of member-expansion. In
the latter module, once we've determined the top-level structure of the object,
we call in to TypeConverter to normalize the properties of the object in normal
mode.

A pleasant side-effect is that we can now remove all member-expansion information
from the environment. The only information we need to pass are two flags:
`include_proto_members` and `idx_hook`, which work the same as before.
These are now passed as flags to the normalizer entry point. The additional
`instance_member_expansion` and `proto` fields that used to be in the
environment are now passed as parameters within ExpandMembersConverter.

Reviewed By: vrama628

Differential Revision: D21657203

fbshipit-source-id: de7bd72aff41517a35a1853d80a6b4997fa8f2d8