Handle unresolved exported identifier in ES6 lint rules
Summary:
The new ES6 lint rules are not properly handling the case when an exported identifier cannot be resolved. We use the `Scope_api.def_of_use` method to find the loc of an identifier's definition, but if the identifier cannot be resolved this function throws, causing flow to crash. (Note this is not causing any problems in production as running these lint rules is gated behind the `experimental.strict_es6_import_export` flag).

Instead, let's add a new `Scope_api.def_of_use_opt` function that optionally returns the definition to avoid crashing on malformed code. We should then use the new `Scope_api.def_of_use_opt` function to resolve export specifiers when checking for the new ES6 lint rules.

Reviewed By: pieterv

Differential Revision: D21832029

fbshipit-source-id: 0492147b7575a4cd2a278359f0443b0996bf021d