Lint rule for use of both toplevel imports and requires
Summary:
The last lint rule we would like to enforce for ES6 imports/exports at the moment is that you cannot mix non-type toplevel  `import` and `require` statements - each file must use either toplevel `imports` or toplevel `requires`. (Type imports are allowed alongside `requires` as there is no alternative CJS syntax for type imports).

This new lint rule is called `mixed-import-and-require` and is gated behind the `experimental.strict_es6_import_export` flowconfig flat to assist with testing and rollout.

We can implement this by detecting imports and requires in the initial pass to gather declarations within the `Strict_es6_import_export` module. We must detect only import statements that contain a value import. Requires are a bit tricker to detect, but we should catch the most common pattern which is either a direct require or a require wrapped in a member expression (`const Foo = require('Foo')` or `const Named = require('Foo').named`). If both an import and require are detected, we display an error that shows the location of both the import and require in the file.

Reviewed By: pieterv

Differential Revision: D21826309

fbshipit-source-id: a6466e72439e8b8cc287822a5dc6ae65322854e6